+++ src/path_stats.ts
@@ -1,0 +1,4 @@
+import { sep, dirname } from 'path';
+
+/** Gets _output_ directory name. */
+export const getBuildDirectory = (url: string) => dirname(url).split(sep)[0];